+++ CHANGELOG.rst
@@ -8,6 +8,7 @@ Release date: ``2019-xx-xx``
 
 - `NXPY-102 <https://jira.nuxeo.com/browse/NXPY-102>`__: Set Upload operations to void operations
 - `NXPY-103 <https://jira.nuxeo.com/browse/NXPY-103>`__: Launch flake8 on actual client data
+- `NXPY-104 <https://jira.nuxeo.com/browse/NXPY-104>`__: Do not log server response based on content length but content type
 
 Technical changes
 -----------------